+++ src/main/java/kr/or/bit/controller/GoogleController.java
@@ -36,7 +36,7 @@ public class GoogleController {
 			// 아이디가 이미 존재하면 바로 메인 페이지로 이동
 			System.out.println("아이디 존재");
 			session.setAttribute("loginuser", loginuser);
-			return "projectList";
+			return "main/main";
 		} else {
 			// 없으면 디비에 insert 후에 이동
 			System.out.println("아이디 없음");
@@ -49,7 +49,7 @@ public class GoogleController {
 			session.setAttribute("loginuser", loginuser);
 
 			userservice.insertUser(u); // mysql에 user insert : 회원가입 성공
-			return "projectList";
+			return "main/main";
 		}
 
 	}
