+++ src/main/java/kr/or/bit/controller/HomeController.java
@@ -1,12 +1,20 @@
 package kr.or.bit.controller;
 
 import java.util.Locale;
+
+import javax.servlet.http.HttpSession;
+
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
+import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestMethod;
+import org.springframework.web.bind.annotation.RequestParam;
+import org.springframework.web.bind.annotation.ResponseBody;
+
+import kr.or.bit.dto.user;
 
 @Controller
 public class HomeController {
@@ -20,10 +28,10 @@ public class HomeController {
 		return "login";
 	}
 	
-	//로그인 성공 후 projectList 페이지로 이동 
-	@RequestMapping(value = "projectList.pie", method = RequestMethod.GET)
+	//로그인 성공 후 projectList 페이지로 이동 	
+	@RequestMapping(value = "main.pie", method = RequestMethod.GET)
 	public String projectList() {
-		return "projectList";
+		return "main/main";
 	}
 	
 	//회원가입 페이지로 이동 
@@ -43,6 +51,16 @@ public class HomeController {
 	public String pwdForgot_emailRequest() {
 		return "user/pwdForgot_emailRequest";
 	}
+	
+	//비밀번호 인증 완료 후 비밀번호 변경 페이지로 이동 
+	@RequestMapping(value = "changePwdAfterCertify.pie", method = RequestMethod.GET)
+	public String changePwdAfterCertify(@RequestParam("email") String email, HttpSession session) {
+		System.out.println("비밀번호 변경 이메일 : "+email);
+		session.setAttribute("email", email);
+		return "user/pwdForgot_reset";
+	}
+	
+	
 	
 	
 }
