+++ src/main/webapp/WEB-INF/spring/root-context.xml
@@ -53,8 +53,8 @@
 		class="org.springframework.mail.javamail.JavaMailSenderImpl">
 		<property name="host" value="smtp.naver.com" />
 		<property name="port" value="465" />
-		<property name="username" value="jgdoh92@naver.com" />
-		<property name="password" value="rhdqngkwk" />
+		<property name="username" value="cavok699@naver.com" />
+		<property name="password" value="1b1j1hbjH@" />
 		<property name="defaultEncoding" value="utf-8" />
 		<property name="javaMailProperties">
 			<props>
