+++ src/main/webapp/WEB-INF/velocity/mail.vm
@@ -25,7 +25,7 @@
                         <table width="100%" cellpadding="0" cellspacing="0" border="0" >
                         <tbody>
                             <tr>
-                                <td width="200" height="200" valign="top"><font color="#006190"><br><a href ="http://localhost:8090/bit/checkEmail.do?check=${ranCheck}">여기</a>를 클릭하시면 회원가입이 완료됩니다!</font></td>
+                                <td width="200" height="200" valign="top"><font color="#006190"><br><a href ="http://localhost:8090/checkEmail.pie?check=${ranCheck}">여기</a>를 클릭하시면 회원가입이 완료됩니다!</font></td>
                                
                             </tr>
